+++ src/leexi/client.ts
@@ -25,9 +25,12 @@ interface ClientOptions {
 }
 
 interface ListParams {
-  since?: string;
-  limit?: number;
+  items?: number;
   page?: number;
+  order?: string;
+  dateFilter?: "created_at" | "performed_at" | "updated_at";
+  from?: string;
+  to?: string;
 }
 
 export interface CallsList {